HiringManagerBackend/requirements.txt
django
django-rest-framework
pillow
PyPDF2
openai
django-cors-headers